feat: rename revert tests args
