+++ zk_toolbox/crates/zk_supervisor/src/commands/test/args/mod.rs
@@ -1,2 +1,2 @@
 pub mod integration;
-pub mod revert_and_restart;
+pub mod revert;
